fix: add BloomPostProcessPass to CMake sources, fix OpaquePass SceneData struct, silence DSP/network warnings
- CMake: add BloomPostProcessPass.cpp to Engine source list (fix link error)
- OpaquePass: adapt to nested SceneData struct, qualify namespace
- BloomPostProcessPass.h: add missing Export.h include
- Audio DSP nodes: add explicit float() casts for M_PI (truncation warnings)
- SpriteAnimation/JobSystem: rename params to avoid name shadowing
- SocketTypes: add WIN32_LEAN_AND_MEAN guard
- UDPTransport: fix platform-conditional variable placement
- ProfilerCPU: suppress unused return warning
- CompilerOptions: add /wd4100 suppression
- miniaudio: suppress MSVC /wd4244
- NeoEditor: add main.cpp entry point

+++ src/engine/audio/dsp/nodes/DelayNode.h
@@ -51,7 +51,7 @@ public:
             for (uint32_t f = 0; f < frames; ++f) {
                 float dry = src[f];
 
-                uint32_t readPos = (m_writePos + m_buffer.size() - chDelay) % m_buffer.size();
+                uint32_t readPos = static_cast<uint32_t>((m_writePos + m_buffer.size() - chDelay) % m_buffer.size());
                 float delayed = m_buffer[readPos];
 
                 m_lpfState = m_lpfState + lpf * (delayed - m_lpfState);
